Added Smart Guardian Tank ("AI" LIKE TANK), All-out-Attack/Defense Strategy mode
If player enters region near enemy castle the guardian tank will get moving from his sedentary postion and go kill the newcoming invader. If player tank fears and leaves guardian tank returns back to his guardian position.
Guardian tanks are available only in level 2+.
If a player is shooting the enemycastle and some tanks are away somewhere in the map, they will "feel :angry:  " . If they are closer to their king they will return to protect him, but if they are closer to the other king they will find the route to go and kill him.